requirements.txt
librosa==0.11.0
pytorch-lightning==2.6.1
lightning==2.6.1
transformers>=4.48.0
moviepy==1.0.3
numpy>=2.0.2
openai>=1.55.3,<2
opencv-python==4.11.0.86
openpyxl==3.1.5
pandas>=2.2.3
pydub==0.25.1
PyYAML==6.0.3
replicate==0.33.0
requests==2.32.5
resampy==0.4.3
spacy==3.8.11
streamlit==1.49.1
streamlit-searchbox
yt-dlp
json-repair
ruamel.yaml
InquirerPy
autocorrect-py
ctranslate2>=4.5.0
edge-tts
pyannote-audio>=4.0.0

# demucs and whisperx are installed separately in install.py
# to avoid torchaudio version conflicts between demucs (<2.2) and whisperx (~=2.8.0).
# Both work fine with torchaudio 2.8+ at runtime (tested up to 2.8.0+cu128).
whisperx>=3.8.1

syllables
pypinyin
g2p-en
xmltodict


fastapi>=0.135.0
sqlmodel>=0.0.37
uvicorn>=0.42.0

